feat: build Particle from dict
