+++ tests/unit/io/conftest.py
@@ -1,0 +1,13 @@
+import pytest
+
+from expertsystem.particle import ParticleCollection
+
+
+@pytest.fixture(scope="session")
+def particle_selection(particle_database: ParticleCollection):
+    selection = ParticleCollection()
+    selection += particle_database.filter(lambda p: p.name.startswith("pi"))
+    selection += particle_database.filter(lambda p: p.name.startswith("K"))
+    selection += particle_database.filter(lambda p: p.name.startswith("D"))
+    selection += particle_database.filter(lambda p: p.name.startswith("J/psi"))
+    return selection
